requirements.txt
django==1.10.3
djangorestframework==3.5.2
django-extensions